The Multiple ECDF's command is incorrect
This guide to creating ECDF's in R was almost very useful to me, but when I tried using the command for multiple ECDF's it only created a single variable ECDF.
The command needs a group="" within the aes() to actually create two or more CDF's, which is what I added to the code. I had to spend some time googling to find this out. Hopefully with this correction the next person will find everything they need on this page.